requirements.txt
azureml-core>=1.15.0
azureml-dataprep
azureml-mlflow
papermill
black
